Gloom
This was a project in MTU - CS 1131. Overall it was my favorite project.
The project was to implement an interpreter for a language that the
instructor had defined. This language, which he called "Gloom", is a
concatentative language similar to factor (I don't know if he has a
language specification or not).
Provided is my implementation of the interpreter and the files and test
cases that our interpreters had to pass.
My implementation differs from the original specification in syntax and
expressiveness, mainly through the addition of "named arguments". I also
attempted to implement an import system for java and gloom files (only
works for gloom files)